iface_forecastio.requirements.txt
argparse==1.2.1
distribute==0.6.24
python-forecastio==1.2
requests==2.3.0
wsgiref==0.1.2
